LayoutTests/http/tests/xmlhttprequest/access-control-preflight-credential-sync-expected.txt

CONSOLE MESSAGE: http://localhost:8000/cookies/resources/cookie-utility.py?queryfunction=setFooCookie
CONSOLE MESSAGE: XHR response - Set the foo cookie
Test case for bug 37781: [XHR] Cross-Origin synchronous request with credential raises NETWORK_ERR

PASSED
